requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==39.0.2
Django==3.2.9
django-crispy-forms==2.0
django-daraja==1.2.0
django-filter==22.1
django-jazzmin==2.6.0
django-phonenumber-field==7.0.2
idna==3.4
phonenumbers==8.13.7
Pillow==9.4.0
pycparser==2.21
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
